// src_bak/utils/types.ts
export type ISchool = {
  id: schoolType;
  description: string;
  name: string;
};

export type IPlayerSpell = ISpell & {
  copy: number;
  selected: boolean;
  owner: "hero" | "enemy";
  updates: IPlayerSpellUpdate[];
  created_at: Date;
};
export type IPlayerResource = IResource & {
  quantity: number;
  created_at: Date;
};
export type INPC = ICharacter & {
  image: string;
  dialogue: string | null;
  created_at: Date;
};
export type IHero = ICharacter & {
  element: IElement;
};
export type IPlayerHero = IHero & { selected: boolean; created_at: Date };

export type IEnemy = ICharacter & {
  element: IElement;
};

export type IEnemyFight = IEnemy & { spells: IPlayerSpell[] };

export type ISpellUpdate = {
  id: string;
  school: ISchool;
  mana: number;
  effect: spellEffectType;
  action: IUpdateAction;
  price: { action: spellUpdatePrice; strength: number | string } | null;
  name: string;
  description: string;
  resource_base: IResource[];
};

export type IPlayerSpellUpdate = ISpellUpdate & {
  created_at: Date;
  resource_base: IResource[];
};

export type IUpdateAction = {
  item: spellUpdateEffect;
  data: number | string;
};
// ...
export type IResource = {
  id: string;
  name: string;
  school: ISchool;
  description: string;
  commonality: number;
};
// ...
export type spellEffectType =
  | "h_heal"
  | "h_trumpremove"
  | "h_trumpset"
  | "h_redraw"
  | "h_mana"
  | "e_redraw"
  | "e_drop"
  | "e_reduice"
  | "h_enforce";

export type herosSelectionError = "none" | "less" | "more" | "incorrect" | null;
export type spellUpdateEffect =
  | "health"
  | "trump"
  | "redraw"
  | "mana"
  | "drop"
  | "strength";
export type spellUpdatePrice = "health";
export type dialogueLayout = "single" | "double" | "triple";
export type schoolType =
  | "restoration"
  | "amplification"
  | "oblation"
  | "alteration"
  | "deception";